re #16, Now busy making a processor to translate the shapes of the tetris to coordinates within the grid

+++ Bright Blocks/Assets/Scripts/MainBlock.cs
@@ -4,7 +4,7 @@ using UnityEngine;
 
 public class MainBlock : Block
 {
-    public float positionOfAttachedBlocks;
+    public int positionOfAttachedBlocks;
 
     private Color currentColorOfSelfAndAttachedBlocks;
     private Color baseColor;
